add higher algorithm to moving entity

+++ src/entity.py
@@ -5,12 +5,12 @@ from src.interface import Animate
 
 class Entity(object):
     status: str
-    x: int
-    y: int
-    w: int
-    h: int
-    range_x: int  # khoảng cách render giữa frame và image
-    range_y: int
+    x: float
+    y: float
+    w: float
+    h: float
+    range_x: float  # khoảng cách render giữa frame và image
+    range_y: float
     speed: float
     action: str
     running: bool
